tests/wpt/tests/css/css-typed-om/the-stylepropertymap/properties/outline-width.html

<meta charset="utf-8">
<title>'outline-width' property</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om-1/#property-stle-value-normalization">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

function assert_is_zero_px(result) {
  assert_style_value_equals(result, new CSSUnitValue(0, 'px'));
}

runPropertyTests('outline-width', [
  // Computed value is 0 when outline-style is 'none'.
  // FIXME: Add separate test where outline-style is not 'none' or 'hidden'.
  {
    syntax: 'thin',
    computed: (_, result) => assert_is_zero_px(result)
  },
  {
    syntax: 'medium',
    computed: (_, result) => assert_is_zero_px(result)
  },
  {
    syntax: 'thick',
    computed: (_, result) => assert_is_zero_px(result)
  },
  {
    syntax: '<length>',
    specified: assert_is_equal_with_range_handling,
    computed: (_, result) => assert_is_zero_px(result)
  },
]);

</script>
